refactor: separate Plane component
